created tank enemies
