+++ Assets/Prefabs/Bullets.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a0007a119bbbe4342afd3d592ad0d597
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
