+++ Assets/Prefabs/Tanks/Blue Tank.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bd82c8d2b7065054abb66d0bbcc52716
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
