+++ Assets/Prefabs/Tanks/Enemy Tanks.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 56fb3d967611f3649be5bb3d13a1e7e8
-NativeFormatImporter:
+guid: d8183cd6429ec1241a4291248f7bc96e
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
