+++ Assets/Prefabs/Tanks/Enemy Tanks/Green tank.prefab.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 2e061bd53c44b8b4db2d0d3232ca45e8
-NativeFormatImporter:
+guid: 9549194a2f8da14459b1f74f6c4bc537
+PrefabImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
