+++ Assets/Prefabs/Tanks/Enemy Tanks/Red Tank.prefab.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 85e74817a3e78974cad3d1f1df28eabc
+guid: 75bf03bc144e0d643a3d5f37988efc25
 PrefabImporter:
   externalObjects: {}
   userData:
