+++ Assets/Prefabs/Tanks/Enemy Tanks/Yellow Tank.prefab.meta
@@ -1,8 +1,7 @@
 fileFormatVersion: 2
-guid: 2e746a4b40cc92949a7e026037df0dee
-NativeFormatImporter:
+guid: 6ec8c26fe24a6f44ea136d93f49fec33
+PrefabImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
