+++ Assets/ScriptableObjects/Enemy Tank.asset
@@ -10,11 +10,9 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 675efd0f0ff087244b6e0f6bddfbbe13, type: 3}
-  m_Name: EnemyTank Red
+  m_Name: Enemy Tank
   m_EditorClassIdentifier: 
-  tankName: Red Enemy
-  color: 2
-  health: 100
-  damage: 10
+  health: 60
+  damage: 15
   movementSpeed: 17
-  rotationSpeed: 500
+  rotationSpeed: 150
