+++ Assets/ScriptableObjects/Enemy Tank.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8124e53abf89c274db6cd88a241455d8
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
