+++ Assets/ScriptableObjects/Player Tank.asset
@@ -10,10 +10,8 @@ MonoBehaviour:
   m_Enabled: 1
   m_EditorHideFlags: 0
   m_Script: {fileID: 11500000, guid: 675efd0f0ff087244b6e0f6bddfbbe13, type: 3}
-  m_Name: PlayerTank green
+  m_Name: Player Tank
   m_EditorClassIdentifier: 
-  tankName: Tank green
-  color: 1
   health: 100
   damage: 20
   movementSpeed: 17
