+++ Assets/ScriptableObjects/Player Tank.asset.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: b5150e5c2e95e0c4e974d4b0a1317a43
+guid: dea38c5f0edd579498ff4e8e95ae13b1
 NativeFormatImporter:
   externalObjects: {}
   mainObjectFileID: 11400000
