--- Assets/ScriptableObjects/PlayerTank Blue.asset
@@ -1,20 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 675efd0f0ff087244b6e0f6bddfbbe13, type: 3}
-  m_Name: PlayerTank Blue
-  m_EditorClassIdentifier: 
-  tankName: player tank blue
-  color: 2
-  health: 100
-  damage: 20
-  movementSpeed: 17
-  rotationSpeed: 150
