+++ Assets/Scripts/MVC/Enemy.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5249dac6e3d9c874b82c6fd95afcf4f3
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
