+++ Assets/Scripts/MVC/Enemy/EnemyTankServices.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ec18f26e949bfb244bfbf1b816cb00d3
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
