+++ Assets/Scripts/MVC/Enemy/EnemyTankView.cs
@@ -1,0 +1,15 @@
+﻿using UnityEngine;
+
+namespace TankBattle.MVC.Enemy
+{
+    public class EnemyTankView : MonoBehaviour
+    {
+        private TankController tankController;
+
+        public void setTankController(TankController tankController)
+        {
+            this.tankController = tankController;
+        }
+
+    }
+}
