+++ Assets/Scripts/MVC/Enemy/EnemyTankView.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 8c25dcf0b63d91a4bab855d99f4da0b5
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
