+++ Assets/Scripts/MVC/Enemy/TankController.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 0b067cf84f42c4a4b87b13b80a98def0
+guid: 0cbd39c3f7693e043ae768eeecb02c60
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
