+++ Assets/Scripts/MVC/Enemy/TankModel.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b67658f05b24d7b47ae26ccc2f28c323
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
