+++ Assets/Scripts/MVC/Player.meta
@@ -1,8 +1,8 @@
 fileFormatVersion: 2
-guid: 22ff1a63ccf1dd44ba3c7aa0adfa9302
-NativeFormatImporter:
+guid: 44a801a07249d154b8c9c471aca95160
+folderAsset: yes
+DefaultImporter:
   externalObjects: {}
-  mainObjectFileID: 11400000
   userData: 
   assetBundleName: 
   assetBundleVariant:
